requirements.txt
checkov>=3.0.0
requests>=2.31.0
fpdf2>=2.7.4
fastapi>=0.103.1
uvicorn>=0.23.2
neo4j>=5.12.0
kafka-python>=2.0.2

rich>=13.0.0

pytest>=7.4.0
